backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
llama-index==0.10.12
llama-index-vector-stores-pinecone==0.1.3
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-groq==0.1.3
pinecone-client==3.1.0
python-dotenv==1.0.1
nest-asyncio==1.6.0
